==> code/1.1 scrape_stats.nba.com.ipynb <==
import pandas as pd
import numpy as np
import requests
import time
from scrapy.selector import Selector
from scrapy.http import HtmlResponse
from selenium import webdriver
from bs4 import BeautifulSoup
import json

chromedriver = "/Users/edwardlee/Desktop/chromedriver"

browser = webdriver.Chrome(executable_path = chromedriver)

# Get all years for 1996-2016
browser.get('http://stats.nba.com/league/player/')
HTML = browser.page_source
time.sleep(5)
year_data = Selector(text=HTML).xpath("//select[@name='Season']/option/text()").extract()
# all_stats = Selector(text=HTML).xpath("//select[@name='section-links']/option/text()").extract()
time.sleep(5)

# Get all years for gamelogs from 1947-2016
browser.get('http://stats.nba.com/league/player/#!/gamelogs/')
HTML_logs = browser.page_source
time.sleep(5)
logs_year = Selector(text=HTML_logs).xpath("//select[@name='Season']/option/text()").extract()
time.sleep(5)

# Get all play types for category Play Type
# browser.get('http://stats.nba.com/league/player/#!/playtype/')
# HTML_play_type = browser.page_source
# play_type_cats = Selector(text=HTML_play_type).xpath("//div[@class='tabs-wrap']/div[@class='tab']/a/text()").extract()

# Remove first array that says "All Season"
logs_year = logs_year[1:]

# Set parameters
general_para  = ['Base', 'Advanced', 'Misc', 'Scoring', 'Usage', 'Opponent', 'Defense']
clutch_para   = ['Base', 'Advanced', 'Misc', 'Scoring', 'Usage']
shoot_para = ['Base', 'Opponent']
playtype_para = ['Transition', 'Isolation', 'PRBallHandler', 'PRRollman', 'Postup', 'Spotup', 'Handoff', 'Cut', 'OffScreen', 'OffRebound', 'Misc']

## Traditional
# Base
trad_base = 'http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Advanced
trad_adv = 'http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Misc
trad_misc = 'http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Scoring
trad_score = 'http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Usage
trad_usage = 'http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Opponent
trad_opp = 'http://stats.nba.com/stats/leagueplayerondetails?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Defense
trad_def = 'http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

## Clutch
# Base
clutch_base = 'http://stats.nba.com/stats/leaguedashplayerclutch?AheadBehind=Ahead+or+Behind&ClutchTime=Last+5+Minutes&College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Per48&Period=4&PlayerExperience=&PlayerPosition=&PlusMinus=N&PointDiff=5&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Advanced
clutch_adv = 'http://stats.nba.com/stats/leaguedashplayerclutch?AheadBehind=Ahead+or+Behind&ClutchTime=Last+5+Minutes&College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Per48&Period=4&PlayerExperience=&PlayerPosition=&PlusMinus=N&PointDiff=5&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Misc
clutch_misc = 'http://stats.nba.com/stats/leaguedashplayerclutch?AheadBehind=Ahead+or+Behind&ClutchTime=Last+5+Minutes&College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Per48&Period=4&PlayerExperience=&PlayerPosition=&PlusMinus=N&PointDiff=5&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Scoring
clutch_score = 'http://stats.nba.com/stats/leaguedashplayerclutch?AheadBehind=Ahead+or+Behind&ClutchTime=Last+5+Minutes&College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Per48&Period=4&PlayerExperience=&PlayerPosition=&PlusMinus=N&PointDiff=5&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Usage
clutch_usage = 'http://stats.nba.com/stats/leaguedashplayerclutch?AheadBehind=Ahead+or+Behind&ClutchTime=Last+5+Minutes&College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Per48&Period=4&PlayerExperience=&PlayerPosition=&PlusMinus=N&PointDiff=5&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

## Shooting
# Shooting Stats
shooting = 'http://stats.nba.com/stats/leaguedashplayershotlocations?College=&Conference=&Country=&DateFrom=&DateTo=&DistanceRange=5ft+Range&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Opponent Shooting Stats
opp_shooting = 'http://stats.nba.com/stats/leaguedashplayershotlocations?College=&Conference=&Country=&DateFrom=&DateTo=&DistanceRange=5ft+Range&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=%s&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

## Player Bios
player_bios = 'http://stats.nba.com/stats/leaguedashplayerbiostats?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&Season=%s&SeasonSegment=&SeasonType=Regular+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

### Tracking Shots (Data fails to load on website)
    
## Game Log, sorted by Date (1946-2015)
game_log = 'http://stats.nba.com/stats/leaguegamelog?Counter=1000&Direction=DESC&LeagueID=00&PlayerOrTeam=P&Season=%s&SeasonType=Regular+Season&Sorter=DATE'

## Play Type (2015-16 only)
play_type = 'http://stats-prod.nba.com/wp-json/statscms/v1/synergy/player/?category=%s&limit=500&name=offensive&q=2456678&season=2015&seasonType=Reg'

## Hustle Stats (2015-16 only)
hustle = 'http://stats.nba.com/stats/leaguehustlestatsplayer?College=&Conference=&Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=Totals&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=2015-16&SeasonSegment=&SeasonType=Playoffs&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight='

# Save all loop-able links into lists to loop later
traditional_stats = [trad_base, trad_adv, trad_misc, trad_score, trad_usage, trad_opp, trad_def]
clutch_stats      = [clutch_base, clutch_adv, clutch_misc, clutch_score, clutch_usage]
shoot_stats       = [shooting, opp_shooting]

browser = webdriver.Chrome(executable_path = chromedriver)
# Traditional Stats, 1996-2016, 7 unique parameters
for trad, stats in zip(traditional_stats, general_para):
    for year in year_data:
        print stats, year
        # Update and save link into variable
        link = trad % (stats, year)
        # Open browser driver through selenium
        browser.get(link)
        # Save selenium page source into variable
        HTML = browser.page_source
        # Save page source into beautifulsoup format
        soup = BeautifulSoup(HTML)
        try:
            # Get all text within <pre> tag and decode from utf-8
            get_text = soup.findAll('pre')[0].getText().encode('utf-8')
            # Load text as json
            json_text = json.loads(get_text)
            # Save json into pandas dataframe
            data = pd.DataFrame(json_text['resultSets'][0]['rowSet'], columns=json_text['resultSets'][0]['headers'])
            # Save to csv
            save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/traditional_stats/%s_%s.csv' % (stats,year)
            data.to_csv(save_csv)
        except:
            print 'Failed', link
        time.sleep(2*np.random.rand())

browser.quit()

browser = webdriver.Chrome(executable_path = chromedriver)
# Clutch Stats, 1996-2016, 5 unique parameters
for clutch, stats in zip(clutch_stats, clutch_para):
    for year in year_data:
        print stats, year
        # Update and save link into variable
        link = clutch % (stats, year)
        # Open browser driver through selenium
        browser.get(link)
        # Save selenium page source into variable
        HTML = browser.page_source
        # Save page source into beautifulsoup format
        soup = BeautifulSoup(HTML)
        try:
            # Get all text within <pre> tag and decode from utf-8
            get_text = soup.findAll('pre')[0].getText().encode('utf-8')
            # Load text as json
            json_text = json.loads(get_text)
            # Save json into pandas dataframe
            data = pd.DataFrame(json_text['resultSets'][0]['rowSet'], columns=json_text['resultSets'][0]['headers'])
            # Save to csv
            save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/clutch_stats/%s_%s_per48.csv' % (stats,year)
            data.to_csv(save_csv)
        except:
            print 'Failed', link
        time.sleep(2*np.random.rand())

browser.quit()

browser = webdriver.Chrome(executable_path = chromedriver)
# Shooting Stats, 1996-2016, 2 unique parameters
for shoot, stats in zip(shoot_stats, shoot_para):
    for year in year_data:
        print stats, year
        # Update and save link into variable
        link = shoot % (stats, year)
        # Open browser driver through selenium
        browser.get(link)
        # Save selenium page source into variable
        HTML = browser.page_source
        # Save page source into beautifulsoup format
        soup = BeautifulSoup(HTML)
        try:
            # Get all text within <pre> tag and decode from utf-8
            get_text = soup.findAll('pre')[0].getText().encode('utf-8')
            # Load text as json
            json_text = json.loads(get_text)
            # Create custom headers for each dataframe
            distance_col = json_text['resultSets']['headers'][0]['columnNames']
            distance_col = [col.replace('.','').replace(' ','_').lower() for col in distance_col]
            tmp = len(json_text['resultSets']['headers'][1]['columnNames'][:5])
            tmp1 = len(json_text['resultSets']['headers'][1]['columnNames'][6:])
            new_cols = []

            for i in range(tmp):
                new_cols.append(json_text['resultSets']['headers'][1]['columnNames'][i])

            for col in distance_col:
                for i in range(tmp, tmp1):
                    save_tmp = json_text['resultSets']['headers'][1]['columnNames'][i]
                    name_replace = save_tmp + '_' + col
                    new_cols.append(name_replace)
                    if save_tmp == 'FG_PCT':
                        break
                    elif save_tmp == 'OPP_FG_PCT':
                        break
            # Save json into pandas dataframe
            data = pd.DataFrame(json_text['resultSets']['rowSet'], columns=new_cols)
            # Save to csv
            save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/shooting_stats/%s_%s.csv' % (stats,year)
            data.to_csv(save_csv)
        except:
            print 'Failed', link

        time.sleep(2*np.random.rand())
browser.quit()

browser = webdriver.Chrome(executable_path = chromedriver)
# Player bios, 1996-2016, no unique parameters
for year in year_data:
    print year
    # Update and save link into variable
    link = player_bios % year
    # Open browser driver through selenium
    browser.get(link)
    # Save selenium page source into variable
    HTML = browser.page_source
    # Save page source into beautifulsoup format
    soup = BeautifulSoup(HTML)
    try:
        # Get all text within <pre> tag and decode from utf-8
        get_text = soup.findAll('pre')[0].getText().encode('utf-8')
        # Load text as json
        json_text = json.loads(get_text)
        # Save json into pandas dataframe
        data = pd.DataFrame(json_text['resultSets'][0]['rowSet'], columns=json_text['resultSets'][0]['headers'])
        # Save to csv
        save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/player_bios/%s_%s.csv' % ('bios',year)
        data.to_csv(save_csv)
    except:
        print 'Failed', link
    time.sleep(2*np.random.rand())

browser.quit()

browser = webdriver.Chrome(executable_path = chromedriver)
# Hustle Stats, 1 season, no unique parameters

# Update and save link into variable
link = hustle
# Open browser driver through selenium
browser.get(link)
# Save selenium page source into variable
HTML = browser.page_source
# Save page source into beautifulsoup format
soup = BeautifulSoup(HTML)
# Get all text within <pre> tag and decode from utf-8
get_text = soup.findAll('pre')[0].getText().encode('utf-8')
# Load text as json
json_text = json.loads(get_text)
# Save json into pandas dataframe
data = pd.DataFrame(json_text['resultSets'][0]['rowSet'], columns=json_text['resultSets'][0]['headers'])
# Save to csv
save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/other_stats/%s_%s.csv' % ('hustle','2015-16')
data.to_csv(save_csv)

browser.quit()

browser = webdriver.Chrome(executable_path = chromedriver)
# Game Logs per Season, 1946-2016, no unique parameters
for year in logs_year:
    print year
    # Update and save link into variable
    link = game_log % year
    # Open browser driver through selenium
    browser.get(link)
    # Save selenium page source into variable
    HTML = browser.page_source
    # Save page source into beautifulsoup format
    soup = BeautifulSoup(HTML)
    try:
        # Get all text within <pre> tag and decode from utf-8
        get_text = soup.findAll('pre')[0].getText().encode('utf-8')
        # Load text as json
        json_text = json.loads(get_text)
        # Save json into pandas dataframe
        data = pd.DataFrame(json_text['resultSets'][0]['rowSet'], columns=json_text['resultSets'][0]['headers'])
        # Save to csv
        save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/gamelogs/%s_%s.csv' % ('gamelogs',year)
        data.to_csv(save_csv)
    except:
        print 'Failed', link
    time.sleep(5*np.random.rand())    

# Scrap Yahoo, ESPN, CBS 2016-17 Preseason Ranking

# ESPN Preseason Ranking
rank_list = []
playername_list = []
for startIndex in range(0, 680, 40):
    url_format = "http://games.espn.com/fba/tools/projections?startIndex=%d" % startIndex
    r = requests.get(url_format)
    
    HTML = r.text
    rank_data = Selector(text=HTML).xpath("//td[@class='playertableData']/text()").extract()
    rank_list += rank_data
    playername_data = Selector(text=HTML).xpath("//td[@class='playertablePlayerName']//a//text()").extract()
    playername_list += playername_data
    time.sleep(2*np.random.rand())

print 'Scrap Completed'

rank_list = filter(lambda a: a != 'RNK', rank_list)

len(rank_list), len(playername_list)

espn_ranking = pd.DataFrame({'ESPNRank_16-17':rank_list, 'PLAYER_NAME': playername_list})

espn_ranking.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2016-17_espn_rank.csv')

# CBS Preseason Ranking
rank_list = []
playername_list = []

url_format = "http://www.cbssports.com/fantasy/basketball/draft/averages?&print_rows=9999"
r = requests.get(url_format)

HTML = r.text
rank_data = Selector(text=HTML).xpath("//tr[@class='row1']/td[1]/text()").extract()
rank_data_2 = Selector(text=HTML).xpath("//tr[@class='row2']/td[1]/text()").extract()
rank_list += rank_data
rank_list += rank_data_2
playername_data = Selector(text=HTML).xpath("//tr[@class='row1']/td[2]/a/text()").extract()
playername_data_2 = Selector(text=HTML).xpath("//tr[@class='row2']/td[2]/a/text()").extract()
playername_list += playername_data
playername_list += playername_data_2
time.sleep(2*np.random.rand())

print 'Scrap Completed'

len(rank_list), len(playername_list)

cbs_ranking = pd.DataFrame({'CBSRank_16-17':rank_list, 'PLAYER_NAME': playername_list})

cbs_ranking['CBSRank_16-17'] = cbs_ranking['CBSRank_16-17'].astype(int)

cbs_ranking = cbs_ranking.sort_values('CBSRank_16-17', ascending=True).reset_index(drop=True)

cbs_ranking.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2016-17_cbs_rank.csv')

# import requests

# url = "http://stats.nba.com/league/team/#!/advanced/"
# headers = {'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/39.0.2171.99 Safari/537.36'}

# with requests.Session() as session:
#     session.headers = headers
#     session.get(url, headers=headers)

#     params = {
#         'DateFrom': '',
#         'DateTo': '',
#         'GameScope': '',
#         'GameSegment': '',
#         'LastNGames': '0',
#         'LeagueID': '00',
#         'Location': '',
#         'MeasureType': 'Advanced',
#         'Month': '0',
#         'OpponentTeamID': '0',
#         'Outcome': '',
#         'PaceAdjust': 'N',
#         'PerMode': 'Totals',
#         'Period': '0',
#         'PlayerExperience': '',
#         'PlayerPosition': '',
#         'PlusMinus': 'N',
#         'Rank': 'N',
#         'Season': '2014-15',
#         'SeasonSegment': '',
#         'SeasonType': 'Regular Season',
#         'StarterBench': '',
#         'VsConference': '',
#         'VsDivision': ''
#     }

#     response = session.get('http://stats.nba.com/stats/leaguedashteamstats', params=params)
#     results = response.json()
#     headers = results['resultSets'][0]['headers']
#     rows = results['resultSets'][0]['rowSet']
#     for row in rows:
#         print(dict(zip(headers, row)))
