==> code/2.1 clean-merge_gamelogs.ipynb <==
import pandas as pd
import numpy as np
import requests
import time
from scrapy.selector import Selector
from scrapy.http import HtmlResponse
from selenium import webdriver
from bs4 import BeautifulSoup
import json

chromedriver = "/Users/edwardlee/Desktop/chromedriver"

browser = webdriver.Chrome(executable_path = chromedriver)

# Get all years for 1996-2016
browser.get('http://stats.nba.com/league/player/')
HTML = browser.page_source
year_data = Selector(text=HTML).xpath("//select[@name='Season']/option/text()").extract()
time.sleep(2)

# Get all years for gamelogs from 1947-2016
browser.get('http://stats.nba.com/league/player/#!/gamelogs/')
HTML_logs = browser.page_source
logs_year = Selector(text=HTML_logs).xpath("//select[@name='Season']/option/text()").extract()
time.sleep(2)

# Remove first array that says "All Season"
logs_year = logs_year[1:]

### MERGE ALL GAME LOGS FROM 1996-97 to 2015-16

all_game_logs = []

for year in year_data:
    csv_file = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/gamelogs/gamelogs_%s.csv' % year
    df = pd.read_csv(csv_file)
    df.drop('Unnamed: 0', axis=1, inplace=True)
    df['SEASON'] = [year for i in range(df.shape[0])]
    all_game_logs.append(df)
    
gamelogs_df = pd.concat(all_game_logs)
save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/datasets/merged/all_gamelogs.csv'
gamelogs_df.to_csv(save_csv)

gamelogs_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_gamelogs.csv')

# Removed unwanted columns and percentages unneeded for analysis
gamelogs_df.drop(['Unnamed: 0', 'SEASON_ID', 'TEAM_ABBREVIATION', 'FG_PCT', 'FG3_PCT', 'FT_PCT', 'VIDEO_AVAILABLE'], axis=1, inplace=True)

gamelogs_df.head(2)

gamelogs_df.info()

# Add in new column with FP calculation
gamelogs_df['FP'] = (gamelogs_df['FGM'] * 1.0) - ((gamelogs_df['FGA']-gamelogs_df['FGM']) * (9./11.)) + (gamelogs_df['FTM'] * 1.0) - \
((gamelogs_df['FTA']-gamelogs_df['FTM']) * (3./1.)) + (gamelogs_df['FG3M'] * 0.5) + (gamelogs_df['PTS'] * 1.0) + (gamelogs_df['REB'] * 1.2) + \
(gamelogs_df['AST'] * 1.5) + (gamelogs_df['STL'] * 2.0) + (gamelogs_df['BLK'] * 2.0) - (gamelogs_df['TOV'] * 1.0)

gamelogs_df.sort_values(['FP'], ascending=False).head(5)

gamelogs_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/gamelogs_clean_withFP.csv')

gamelogs_target = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/gamelogs_clean_withFP.csv')

# Groupby by season, player id, player name and fantasy points with agg function of mean, std and var
gamelogs_mean_fp = gamelogs_target.groupby(['SEASON','PLAYER_ID','PLAYER_NAME'])[['FP']].mean().reset_index().rename(columns={'FP':'fp_mean'})
gamelogs_std_fp = gamelogs_target.groupby(['SEASON','PLAYER_ID','PLAYER_NAME'])[['FP']].std().reset_index().rename(columns={'FP':'fp_std'})
gamelogs_var_fp = gamelogs_target.groupby(['SEASON','PLAYER_ID','PLAYER_NAME'])[['FP']].var().reset_index().rename(columns={'FP':'fp_var'})

# Groupby by season, team, fantasy points with agg function of mean, std, var
gamelogs_team_mean_fp = gamelogs_target.groupby(['SEASON', 'TEAM_NAME', 'GAME_DATE'])[['FP']].mean().reset_index().rename(columns={'FP':'fp_team_mean'})

# Name is only indicator, so don't need that for now, can remove
gamelogs_std_fp = gamelogs_std_fp[['PLAYER_ID','SEASON','fp_std']]
gamelogs_var_fp = gamelogs_var_fp[['PLAYER_ID','SEASON','fp_var']]

# Merge mean, std and var together
gamelogs_target_final = gamelogs_mean_fp.merge(gamelogs_std_fp, how='outer', on=['PLAYER_ID', 'SEASON']).merge(gamelogs_var_fp, how='outer', on=['PLAYER_ID', 'SEASON'])

gamelogs_target_final.head()

gamelogs_team_mean_fp.head()

# Save dummy target variables together
gamelogs_target_final.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/gamelogs_fp_targets.csv')

# Save team fp mean
gamelogs_team_mean_fp.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/gamelogs_fp_team_mean.csv')
