==> code/2.3 clean-merge_traditional-shooting-bios-part1.ipynb <==
import pandas as pd
import numpy as np
import requests
import time
from scrapy.selector import Selector
from scrapy.http import HtmlResponse
from selenium import webdriver
from bs4 import BeautifulSoup
import json

chromedriver = "/Users/edwardlee/Desktop/chromedriver"

browser = webdriver.Chrome(executable_path = chromedriver)

# Get all years for 1996-2016
browser.get('http://stats.nba.com/league/player/')
HTML = browser.page_source
year_data = Selector(text=HTML).xpath("//select[@name='Season']/option/text()").extract()
time.sleep(2)

# Set parameters
general_para  = ['Base', 'Advanced', 'Misc', 'Scoring', 'Usage', 'Opponent', 'Defense']
shoot_para = ['Base', 'Opponent']
# playtype_para = ['Transition', 'Isolation', 'PRBallHandler', 'PRRollman', 'Postup', 'Spotup', 'Handoff', 'Cut', 'OffScreen', 'OffRebound', 'Misc', 'hustle']

### MERGE ALL PLAYER BIOS FROM 1996-97 to 2015-16

all_bios_df = []

for year in year_data:
    csv_file = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/player_bios/bios_%s.csv' % year
    df = pd.read_csv(csv_file)
    df.drop('Unnamed: 0', axis=1, inplace=True)
    df['SEASON'] = [year for i in range(df.shape[0])]
    all_bios_df.append(df)

bios_df = pd.concat(all_bios_df)
save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_bios.csv'
bios_df.to_csv(save_csv)

### MERGE ALL SHOOTING STATS FROM 1996-97 to 2015-16

shooting = []
opp_shooting = []

for stats in shoot_para:
    for year in year_data:
        csv_file = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/shooting_stats/%s_%s.csv' % (stats, year)
        df = pd.read_csv(csv_file)
        df.drop('Unnamed: 0', axis=1, inplace=True)
        df['SEASON'] = [year for i in range(df.shape[0])]
        if stats == 'Base':
            shooting.append(df)
        elif stats == 'Opponent':
            opp_shooting.append(df)

shoot_df = pd.concat(shooting)
opp_shoot_df = pd.concat(opp_shooting)
save_csv = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_shoot.csv'
save_csv1 = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_opp_shoot.csv'
shoot_df.to_csv(save_csv)
opp_shoot_df.to_csv(save_csv1)

shoot_df.columns

### MERGE TRADITIONAL STATS BY CATEGORIES FROM 1996-97 to 2015-16

base     = []
advanced = []
misc     = []
scoring  = []
usage    = []
opponent = []
defense  = []

for stats in general_para:
    for year in year_data:
        csv_file = '/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/traditional_stats/%s_%s.csv' % (stats, year)
        df = pd.read_csv(csv_file)
        df.drop('Unnamed: 0', axis=1, inplace=True)
        df['SEASON'] = [year for i in range(df.shape[0])]
        df['win_share'] = 
        if stats == 'Base':
            base.append(df)
        elif stats == 'Advanced':
            advanced.append(df)
        elif stats == 'Misc':
            misc.append(df)
        elif stats == 'Scoring':
            scoring.append(df)
        elif stats == 'Usage':
            usage.append(df)
        elif stats == 'Opponent':
            opponent.append(df)
        elif stats == 'Defense':
            defense.append(df)

base_df     = pd.concat(base)
advanced_df = pd.concat(advanced)
misc_df     = pd.concat(misc)
scoring_df  = pd.concat(scoring)
usage_df    = pd.concat(usage)
opponent_df = pd.concat(opponent)
defense_df  = pd.concat(defense)
base_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_base.csv')
advanced_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_advanced.csv')
misc_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_misc.csv')
scoring_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_scoring.csv')
usage_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_usage.csv')
opponent_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_opponent.csv')
defense_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_defense.csv')

### Merge Player Shooting and Opponent's Shooting
shoot_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_shoot.csv')
opp_shoot_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_opp_shoot.csv')

# Want these cols to merge into player shooting dataset
opponent_col = ['PLAYER_ID', u'OPP_FGM_less_than_5_ft',
       u'OPP_FGA_less_than_5_ft', u'OPP_FG_PCT_less_than_5_ft',
       u'OPP_FGM_5-9_ft', u'OPP_FGA_5-9_ft', u'OPP_FG_PCT_5-9_ft',
       u'OPP_FGM_10-14_ft', u'OPP_FGA_10-14_ft', u'OPP_FG_PCT_10-14_ft',
       u'OPP_FGM_15-19_ft', u'OPP_FGA_15-19_ft', u'OPP_FG_PCT_15-19_ft',
       u'OPP_FGM_20-24_ft', u'OPP_FGA_20-24_ft', u'OPP_FG_PCT_20-24_ft',
       u'OPP_FGM_25-29_ft', u'OPP_FGA_25-29_ft', u'OPP_FG_PCT_25-29_ft',
       u'OPP_FGM_30-34_ft', u'OPP_FGA_30-34_ft', u'OPP_FG_PCT_30-34_ft',
       u'OPP_FGM_35-39_ft', u'OPP_FGA_35-39_ft', u'OPP_FG_PCT_35-39_ft',
       u'OPP_FGM_40+_ft', u'OPP_FGA_40+_ft', u'OPP_FG_PCT_40+_ft', 'SEASON']

opp_shoot_df = opp_shoot_df[wanted_col]

# Merge player shooting and opponent's shooting against the player together and save into csv
complete_shoot = pd.merge(shoot_df, opp_shoot_df, how='inner', on=['PLAYER_ID', 'SEASON'])
complete_shoot.drop(['Unnamed: 0', 'TEAM_ID'], axis=1, inplace=True)
complete_shoot.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/merged_shooting.csv')

trad_base_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_base.csv')
trad_adv_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_advanced.csv')
trad_def_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_defense.csv')
trad_misc_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_misc.csv')
trad_opp_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_opponent.csv')
trad_score_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_scoring.csv')
trad_usage_df = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/merged/all_traditional_usage.csv')

trad_base_df.columns

trad_adv_df.columns

trad_def_df.columns

trad_misc_df.columns

trad_opp_df.columns

trad_score_df.columns

trad_usage_df.columns

trad_base_cols = [u'PLAYER_ID', u'PLAYER_NAME',
       u'TEAM_ABBREVIATION', u'AGE', u'GP', u'W', u'L', u'MIN',
       u'FGM', u'FGA', u'FG3M', u'FG3A', u'FTM', u'FTA',
       u'OREB', u'DREB', u'REB', u'AST', u'TOV', u'STL', u'BLK',
       u'BLKA', u'PF', u'PFD', u'PTS', u'PLUS_MINUS', u'DD2', u'TD3', 'SEASON']

trad_adv_cols = [u'PLAYER_ID', u'OFF_RATING', u'DEF_RATING', u'NET_RATING', u'AST_PCT', u'AST_TO',
       u'AST_RATIO', u'OREB_PCT', u'DREB_PCT', u'REB_PCT', u'TM_TOV_PCT',
       u'EFG_PCT', u'TS_PCT', u'USG_PCT', u'PACE', u'PIE', u'FGM_PG', 'SEASON']

# All columns are in other columns except DEF_WS and DEF_WS is not a good predictor because the equation
# is weighted by the team and not by the player's performance.
# This dataset is completed removed
# trad_def_cols = [u'PLAYER_ID', u'DEF_WS', 'SEASON']

trad_misc_cols = [u'PLAYER_ID', u'PTS_OFF_TOV', u'PTS_2ND_CHANCE', u'PTS_FB', u'PTS_PAINT',
       u'OPP_PTS_OFF_TOV', u'OPP_PTS_2ND_CHANCE', u'OPP_PTS_FB',
       u'OPP_PTS_PAINT', 'SEASON']

# trad_opp_cols = [u'VS_PLAYER_ID', u'OPP_FGM', u'OPP_FGA',
#        u'OPP_FG_PCT', u'OPP_FG3M', u'OPP_FG3A', u'OPP_FG3_PCT', u'OPP_FTM',
#        u'OPP_FTA', u'OPP_FT_PCT', u'OPP_OREB', u'OPP_DREB', u'OPP_REB',
#        u'OPP_AST', u'OPP_TOV', u'OPP_STL', u'OPP_BLK', u'OPP_BLKA', u'OPP_PF',
#        u'OPP_PFD', u'OPP_PTS', 'SEASON']

trad_score_cols = [u'PLAYER_ID', u'PCT_FGA_2PT', u'PCT_FGA_3PT', u'PCT_PTS_2PT', u'PCT_PTS_2PT_MR',
       u'PCT_PTS_3PT', u'PCT_PTS_FB', u'PCT_PTS_FT', u'PCT_PTS_OFF_TOV',
       u'PCT_PTS_PAINT', u'PCT_AST_2PM', u'PCT_UAST_2PM', u'PCT_AST_3PM',
       u'PCT_UAST_3PM', u'PCT_AST_FGM', u'PCT_UAST_FGM', 'SEASON']

trad_usage_cols = [u'PLAYER_ID', u'PCT_FGM', u'PCT_FGA', u'PCT_FG3M', u'PCT_FG3A',
       u'PCT_FTM', u'PCT_FTA', u'PCT_OREB', u'PCT_DREB', u'PCT_REB',
       u'PCT_AST', u'PCT_TOV', u'PCT_STL', u'PCT_BLK', u'PCT_BLKA', u'PCT_PF',
       u'PCT_PFD', u'PCT_PTS', 'SEASON']

shooting_cols = [u'PLAYER_ID', u'FGM_less_than_5_ft',
       u'FGA_less_than_5_ft', u'FG_PCT_less_than_5_ft', u'FGM_5-9_ft',
       u'FGA_5-9_ft', u'FG_PCT_5-9_ft', u'FGM_10-14_ft', u'FGA_10-14_ft',
       u'FG_PCT_10-14_ft', u'FGM_15-19_ft', u'FGA_15-19_ft',
       u'FG_PCT_15-19_ft', u'FGM_20-24_ft', u'FGA_20-24_ft',
       u'FG_PCT_20-24_ft', u'FGM_25-29_ft', u'FGA_25-29_ft',
       u'FG_PCT_25-29_ft', u'FGM_30-34_ft', u'FGA_30-34_ft',
       u'FG_PCT_30-34_ft', u'FGM_35-39_ft', u'FGA_35-39_ft',
       u'FG_PCT_35-39_ft', u'FGM_40+_ft', u'FGA_40+_ft', u'FG_PCT_40+_ft',
       u'SEASON']

opp_shooting_cols = ['PLAYER_ID', u'OPP_FGM_less_than_5_ft',
       u'OPP_FGA_less_than_5_ft', u'OPP_FG_PCT_less_than_5_ft',
       u'OPP_FGM_5-9_ft', u'OPP_FGA_5-9_ft', u'OPP_FG_PCT_5-9_ft',
       u'OPP_FGM_10-14_ft', u'OPP_FGA_10-14_ft', u'OPP_FG_PCT_10-14_ft',
       u'OPP_FGM_15-19_ft', u'OPP_FGA_15-19_ft', u'OPP_FG_PCT_15-19_ft',
       u'OPP_FGM_20-24_ft', u'OPP_FGA_20-24_ft', u'OPP_FG_PCT_20-24_ft',
       u'OPP_FGM_25-29_ft', u'OPP_FGA_25-29_ft', u'OPP_FG_PCT_25-29_ft',
       u'OPP_FGM_30-34_ft', u'OPP_FGA_30-34_ft', u'OPP_FG_PCT_30-34_ft',
       u'OPP_FGM_35-39_ft', u'OPP_FGA_35-39_ft', u'OPP_FG_PCT_35-39_ft',
       u'OPP_FGM_40+_ft', u'OPP_FGA_40+_ft', u'OPP_FG_PCT_40+_ft', 'SEASON']

# Update all previous dfs with similar column statistics and save them into a new df
new_trad_base_df  = trad_base_df[trad_base_cols]
new_trad_adv_df   = trad_adv_df[trad_adv_cols]
# new_trad_def_df   = trad_def_df[trad_def_cols]
new_trad_misc_df  = trad_misc_df[trad_misc_cols]
# new_trad_opp_df   = trad_opp_df[trad_opp_cols]
new_trad_score_df = trad_score_df[trad_score_cols]
new_trad_usage_df = trad_usage_df[trad_usage_cols]

new_trad_opp_df.rename(columns={'VS_PLAYER_ID':'PLAYER_ID'}, inplace=True)

# Merge all columns for traditional stats together by PLAYER_ID
final_trad_df = new_trad_base_df.merge(new_trad_adv_df, how='outer', on=['PLAYER_ID', 'SEASON']) \
#                 .merge(new_trad_def_df, how='outer', on=['PLAYER_ID', 'SEASON']) \
                .merge(new_trad_misc_df, how='outer', on=['PLAYER_ID', 'SEASON']) \
                .merge(new_trad_opp_df, how='outer', on=['PLAYER_ID', 'SEASON']) \
                .merge(new_trad_score_df, how='outer', on=['PLAYER_ID', 'SEASON']) \
                .merge(new_trad_usage_df, how='outer', on=['PLAYER_ID', 'SEASON'])

final_trad_df.shape

final_trad_df.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/final_trad_all.csv')
