==> code/2.5 clean-merge_final.ipynb <==
import pandas as pd
import numpy as np

df_clutch_target = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/clutch_rank_target.csv')
df_trad_all = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/trad_rank_target.csv')
df_gamelogs_target = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/gamelogs_fp_targets.csv')

df_clutch_target.rename(columns={'agg_rank':'clutch_agg_rank'}, inplace=True)

df_clutch_target_merge = df_clutch_target[['PLAYER_ID', 'SEASON', 'clutch_agg_rank']]

df_gamelogs_target_merge = df_gamelogs_target[['PLAYER_ID', 'SEASON', 'fp_mean', 'fp_std', 'fp_var']]

df_complete = df_trad_all.merge(df_clutch_target_merge, on=['PLAYER_ID', 'SEASON'], how='outer').merge(df_gamelogs_target_merge, on=['PLAYER_ID', 'SEASON'], how='outer')

df_complete = df_complete.drop(['Unnamed: 0', 'Unnamed: 0.1'], axis=1).sort_values('agg_rank', ascending=True)

df_complete_target = df_complete[['PLAYER_NAME', 'TEAM_ABBREVIATION', 'SEASON', 'AGE', 'agg_rank', 'clutch_agg_rank', 'fp_mean', 'fp_std', 'fp_var','game_risk']].sort_values('fp_mean', ascending=False)

df_complete.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/complete_all.csv')

df_complete_target.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/completed/complete_all_target.csv')
