==> code/5.3 Results - Final Comparison.ipynb <==
import pandas as pd
import numpy as np

pd.options.display.max_rows = 200

my_rank_2015 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2015.csv', index_col=False)
my_rank_2014 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2014.csv', index_col=False)
my_rank_2013 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2013.csv', index_col=False)
my_rank_2012 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2012.csv', index_col=False)
compare_rank_2015_16 = pd.read_excel('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2015-16_ranks_adp.xls')
compare_rank_2014_15 = pd.read_excel('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2014-15_ranks_adp.xls')
compare_rank_2013_14 = pd.read_excel('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2013-14_rank_adp.xls')
beat_them_all = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2016-17_beat_them.csv')

# ========================================================
# Comparison to Yahoo, ESPN, CBS and BasketballMonster Ranks
# ========================================================

# How to compare:
1. 2012-13 predictions to 2013-14 final ranks
2. 2013-14 predictions to 2014-15 final ranks
3. 2014-15 predictions to 2015-16 final ranks
4. 2015-16 predictions to 2016-17 preseason ranks

compare_rank_2013_14.head(12)

compare_rank_2013_14[['Round', 'Rank', 'Name']].head(12)

my_rank_2013.head(12)

compare_rank_2012_13.head()

my_rank_2011.drop('Unnamed: 0', axis=1, inplace=True)
my_rank_2012.drop('Unnamed: 0', axis=1, inplace=True)
my_rank_2013.drop('Unnamed: 0', axis=1, inplace=True)
my_rank_2014.drop('Unnamed: 0', axis=1, inplace=True)
my_rank_2015.drop('Unnamed: 0', axis=1, inplace=True)

def switcharoo(x):
    if x == 0.0:
        return 'Low Risk'
    elif x == 1.0:
        return 'Caution'
    elif x == 2.0:
        return 'High Risk'
    elif x == 3.0:
        return 'Danger'

my_rank_2011['MyRank'] = [x+1 for x in range(my_rank_2011.shape[0])]
my_rank_2011['agg_rank_score'] = my_rank_2011['agg_rank_score'] * -1
my_rank_2011['clutch_rank_score'] = my_rank_2011['clutch_rank_score'] * -1
my_rank_2011['game_risk_score'] = my_rank_2011['game_risk_score'] * -1
my_rank_2011['game_risk_score'] = my_rank_2011['game_risk_score'].apply(switcharoo)

my_rank_2012['MyRank'] = [x+1 for x in range(my_rank_2012.shape[0])]
my_rank_2012['agg_rank_score'] = my_rank_2012['agg_rank_score'] * -1
my_rank_2012['clutch_rank_score'] = my_rank_2012['clutch_rank_score'] * -1
my_rank_2012['game_risk_score'] = my_rank_2012['game_risk_score'] * -1
my_rank_2012['game_risk_score'] = my_rank_2012['game_risk_score'].apply(switcharoo)

my_rank_2013['MyRank'] = [x+1 for x in range(my_rank_2013.shape[0])]
my_rank_2013['agg_rank_score'] = my_rank_2013['agg_rank_score'] * -1
my_rank_2013['clutch_rank_score'] = my_rank_2013['clutch_rank_score'] * -1
my_rank_2013['game_risk_score'] = my_rank_2013['game_risk_score'] * -1
my_rank_2013['game_risk_score'] = my_rank_2013['game_risk_score'].apply(switcharoo)

my_rank_2014['MyRank'] = [x+1 for x in range(my_rank_2014.shape[0])]
my_rank_2014['agg_rank_score'] = my_rank_2014['agg_rank_score'] * -1
my_rank_2014['clutch_rank_score'] = my_rank_2014['clutch_rank_score'] * -1
my_rank_2014['game_risk_score'] = my_rank_2014['game_risk_score'] * -1
my_rank_2014['game_risk_score'] = my_rank_2014['game_risk_score'].apply(switcharoo)

my_rank_2015['MyRank'] = [x+1 for x in range(my_rank_2015.shape[0])]
my_rank_2015['agg_rank_score'] = my_rank_2015['agg_rank_score'] * -1
my_rank_2015['clutch_rank_score'] = my_rank_2015['clutch_rank_score'] * -1
my_rank_2015['game_risk_score'] = my_rank_2015['game_risk_score'] * -1
my_rank_2015['game_risk_score'] = my_rank_2015['game_risk_score'].apply(switcharoo)

compare_rank_2012_13.drop('CBSRank', axis=1, inplace=True)
compare_rank_2013_14.drop('CBSRank', axis=1, inplace=True)
beat_them_all.drop('Unnamed: 0', axis=1, inplace=True)

compare_rank_2012_13.rename(columns={'Name':'PLAYER_NAME', 'Rank':'BBMRank', 'Y!ORank':'YahooRank'}, inplace=True)
compare_rank_2012_13 = compare_rank_2012_13[['PLAYER_NAME', 'BBMRank', 'YahooRank', 'ESPNRank']]

compare_rank_2013_14.rename(columns={'Name':'PLAYER_NAME', 'Rank':'BBMRank', 'Y!ORank':'YahooRank'}, inplace=True)
compare_rank_2013_14 = compare_rank_2013_14[['PLAYER_NAME', 'BBMRank', 'YahooRank', 'ESPNRank']]

compare_rank_2014_15.rename(columns={'Name':'PLAYER_NAME', 'Rank':'BBMRank', 'Y!ORank':'YahooRank'}, inplace=True)
compare_rank_2014_15 = compare_rank_2014_15[['PLAYER_NAME', 'BBMRank', 'CBSRank', 'YahooRank', 'ESPNRank']]

compare_rank_2015_16.rename(columns={'Name':'PLAYER_NAME', 'Rank':'BBMRank', 'Y!ORank':'YahooRank'}, inplace=True)
compare_rank_2015_16 = compare_rank_2015_16[['PLAYER_NAME', 'BBMRank', 'CBSRank', 'YahooRank', 'ESPNRank']]

my_rank_2011 = my_rank_2011.merge(compare_rank_2012_13, on='PLAYER_NAME', how='left')
my_rank_2012 = my_rank_2012.merge(compare_rank_2013_14, on='PLAYER_NAME', how='left')
my_rank_2013 = my_rank_2013.merge(compare_rank_2014_15, on='PLAYER_NAME', how='left')
my_rank_2014 = my_rank_2014.merge(compare_rank_2015_16, on='PLAYER_NAME', how='left')
my_rank_2015 = my_rank_2015.merge(beat_them_all, on='PLAYER_NAME', how='left')

my_rank_2015.columns

my_rank_2015[['PLAYER_NAME', 'MyRank', 'YahooRank_16_17', 'CBSRank_16-17', 'ESPNRank_16-17']].head(15)

round1 = []
round2 = []
round3 = []
round4 = []
round5 = []
round6 = []
round7 = []
round8 = []
round9 = []
round10 = []
round11 = []
round12 = []
round13 = []
round14 = []

for rank in my_rank_2015['MyRank']:
    if rank <= 12:
        round1.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 12 and rank <= 24:
        round2.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 24 and rank <= 36:
        round3.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 36 and rank <= 48:
        round4.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 48 and rank <= 60:
        round5.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 60 and rank <= 72:
        round6.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 72 and rank <= 84:
        round7.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 84 and rank <= 96:
        round8.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 96 and rank <= 108:
        round9.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 108 and rank <= 120:
        round10.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 120 and rank <= 132:
        round11.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 132 and rank <= 144:
        round12.append(my_rank_2015['PLAYER_NAME'][rank-1])
    elif rank > 144 and rank <= 156:
        round13.append(my_rank_2015['PLAYER_NAME'][rank-1])
    else:
        round14.append(my_rank_2015['PLAYER_NAME'][rank-1])

round1

round2

round3

round4

round5

round6

round7

ranked_dictionary = {
    'r1':round1,
    'r2':round2,
    'r3':round3,
    'r4':round4,
    'r5':round5,
    'r6':round6,
    'r7':round7,
    'r8':round8,
    'r9':round9,
    'r10':round10,
    'r11':round11,
    'r12':round12,
    'r13':round13,
}

tmp = pd.DataFrame(ranked_dictionary)

tmp[['r1', 'r2', 'r3', 'r4', 'r5', 'r6', 'r7', 'r8', 'r9', 'r10', 'r11', 'r12', 'r13']]
