==> code/5.4 Results - Final Comparison_part2.ipynb <==
import pandas as pd
import numpy as np

pd.options.display.max_rows = 200
pd.options.display.float_format = '{:.5f}'.format

my_rank_2015 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2015.csv', index_col=False)
my_rank_2014 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2014.csv', index_col=False)
my_rank_2013 = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/my_rank_2013.csv', index_col=False)
compare_rank_2015_16 = pd.read_excel('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2015-16_ranks_adp.xls')
compare_rank_2014_15 = pd.read_excel('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2014-15_ranks_adp.xls')
compare_rank_2013_14 = pd.read_excel('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2013-14_rank_adp.xls')
beat_them_all = pd.read_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/2016-17_beat_them.csv')

my_rank_2015.drop('Unnamed: 0', axis=1,inplace=True)
my_rank_2014.drop('Unnamed: 0', axis=1,inplace=True)
my_rank_2013.drop('Unnamed: 0', axis=1,inplace=True)

for col in my_rank_2013.columns:
    if col != 'PLAYER_NAME' and col != 'game_risk_score':
        new_col_name = 'rank_%s' % col
        my_rank_2013.sort_values(col, ascending=False, inplace=True)
        my_rank_2013[new_col_name] = [i+1 for i in range(my_rank_2013.shape[0])]

my_rank_2013['final_rank'] = my_rank_2013.ix[:,'rank_fp_mean_score':].mean(axis=1)

for col in my_rank_2014.columns:
    if col != 'PLAYER_NAME' and col != 'game_risk_score':
        new_col_name = 'rank_%s' % col
        my_rank_2014.sort_values(col, ascending=False, inplace=True)
        my_rank_2014[new_col_name] = [i+1 for i in range(my_rank_2014.shape[0])]

my_rank_2014['final_rank'] = my_rank_2014.ix[:,'rank_fp_mean_score':].mean(axis=1)

for col in my_rank_2015.columns:
    if col != 'PLAYER_NAME' and col != 'game_risk_score':
        new_col_name = 'rank_%s' % col
        my_rank_2015.sort_values(col, ascending=False, inplace=True)
        my_rank_2015[new_col_name] = [i+1 for i in range(my_rank_2015.shape[0])]

my_rank_2015['final_rank'] = my_rank_2015.ix[:,'rank_fp_mean_score':].mean(axis=1)

my_rank_2013['game_risk_score'] = my_rank_2013['game_risk_score'] * -1
my_rank_2014['game_risk_score'] = my_rank_2014['game_risk_score'] * -1
my_rank_2015['game_risk_score'] = my_rank_2015['game_risk_score'] * -1

for i in range(my_rank_2013.shape[0]):
    if my_rank_2013['game_risk_score'][i] == 1.0:
        my_rank_2013['final_rank'][i] *= 1.25
    elif my_rank_2013['game_risk_score'][i] == 2.0:
        my_rank_2013['final_rank'][i] *= 1.50
    elif my_rank_2013['game_risk_score'][i] == 3.0:
        my_rank_2013['final_rank'][i] *= 2.00
        
for i in range(my_rank_2014.shape[0]):
    if my_rank_2014['game_risk_score'][i] == 1.0:
        my_rank_2014['final_rank'][i] *= 1.25
    elif my_rank_2014['game_risk_score'][i] == 2.0:
        my_rank_2014['final_rank'][i] *= 1.50
    elif my_rank_2014['game_risk_score'][i] == 3.0:
        my_rank_2014['final_rank'][i] *= 2.00
        
for i in range(my_rank_2015.shape[0]):
    if my_rank_2015['game_risk_score'][i] == 1.0:
        my_rank_2015['final_rank'][i] *= 1.25
    elif my_rank_2015['game_risk_score'][i] == 2.0:
        my_rank_2015['final_rank'][i] *= 1.50
    elif my_rank_2015['game_risk_score'][i] == 3.0:
        my_rank_2015['final_rank'][i] *= 2.00

my_rank_2013[['PLAYER_NAME', 'final_rank']].sort_values('final_rank', ascending=True).reset_index(drop=True).head(36)

my_rank_2014[['PLAYER_NAME', 'final_rank']].sort_values('final_rank', ascending=True).reset_index(drop=True).head(36)

my_rank_2015[['PLAYER_NAME', 'final_rank']].sort_values('final_rank', ascending=True).reset_index(drop=True).head(36)

compare_rank_2013_14[['ESPNRank', 'Name']].sort_values('ESPNRank', ascending=True).head(36)

compare_rank_2014_15[compare_rank_2014_15['CBSRank'] != 0][['CBSRank', 'Name']].sort_values('CBSRank', ascending=True).head(36)

compare_rank_2015_16[compare_rank_2015_16.Rank != 0][['Rank', 'Name']].sort_values('Rank', ascending=True).head(36)

beat_them_all.columns

beat_them_all[['PLAYER_NAME', 'ESPNRank_16-17']].sort_values('ESPNRank_16-17', ascending=True).head(36)

my_rank_2015[['PLAYER_NAME', 'final_rank']].sort_values('final_rank').head(36)

my_rank_2015_final = my_rank_2015[['PLAYER_NAME', 'final_rank']].sort_values('final_rank').reset_index(drop=True)

my_rank_2015_final = my_rank_2015_final[(my_rank_2015_final.PLAYER_NAME != 'Kobe Bryant') & (my_rank_2015_final.PLAYER_NAME != 'Tim Duncan')].reset_index(drop=True)

for i in range(my_rank_2015_final.shape[0]):
    my_rank_2015_final['ranked'][i] = i+1

round1 = []
round2 = []
round3 = []
round4 = []
round5 = []
round6 = []
round7 = []
round8 = []
round9 = []
round10 = []
round11 = []
round12 = []
round13 = []
round14 = []

len(my_rank_2015_final.index)

for rank in my_rank_2015_final['ranked']:
    if rank <= 12:
        round1.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 12 and rank <= 24:
        round2.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 24 and rank <= 36:
        round3.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 36 and rank <= 48:
        round4.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 48 and rank <= 60:
        round5.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 60 and rank <= 72:
        round6.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 72 and rank <= 84:
        round7.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 84 and rank <= 96:
        round8.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 96 and rank <= 108:
        round9.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 108 and rank <= 120:
        round10.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 120 and rank <= 132:
        round11.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 132 and rank <= 144:
        round12.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    elif rank > 144 and rank <= 156:
        round13.append(my_rank_2015_final['PLAYER_NAME'][rank-1])
    else:
        round14.append(my_rank_2015_final['PLAYER_NAME'][rank-1])

ranked_dictionary = {
    'r1':round1,
    'r2':round2,
    'r3':round3,
    'r4':round4,
    'r5':round5,
    'r6':round6,
    'r7':round7,
    'r8':round8,
    'r9':round9,
    'r10':round10,
    'r11':round11,
    'r12':round12,
    'r13':round13,
}

prediction_2016_17 = pd.DataFrame(ranked_dictionary)

prediction_2016_17[['r1', 'r2', 'r3', 'r4', 'r5', 'r6', 'r7', 'r8', 'r9', 'r10', 'r11', 'r12', 'r13']]

prediction_2016_17.to_csv('/Users/edwardlee/Desktop/PROJECTS/NBA Fantasy Project/rank_project/comparison_dataset/prediction_2016_17.csv')
